Fix option menu
