--- Assets/Scenes/OptionMenu.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 2cda990e2423bbf4892e6590ba056729
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
