--- Assets/Scripts/OptionsMenu/OptionsMenu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 140003e98bbf6f94ab08b7d94e590ab1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
